Added leaderboardcard.jsx
